+++ src/main/java/code/clusteringComponent/IncrementalClusterController.java
@@ -23,7 +23,7 @@ public class IncrementalClusterController {
         HashMap<String,Integer> updatedClusterId = hierarchicalClusterer.clusterIncrementally(hashMap);
         List<ClusterInfo> existingInfo = DBConnect.getInstance().getClusterInfo();
         List<ClusterInfo> info = new ClusterInfoHelper().incrementDiameters(updatedClusterId, hashMap, existingInfo);
-        HashMap<String,Integer> clusterRank = new ClusterInfoHelper().incrementalRanking(updatedClusterId,hashMap);
+        HashMap<String,Double> clusterRank = new ClusterInfoHelper().incrementalRanking(updatedClusterId,hashMap);
         DBConnect.getInstance().updateClusterRank(clusterRank);
         DBConnect.getInstance().updateClusterInfo(info);
         DBConnect.getInstance().updateClusterIDs(updatedClusterId);
